edited tests so they all run together from run_tests.py. Added Franco Peschiera to AUTHORS.
